glibc linaro 2.20: only symlink aarch64_be linker if it exists
This fixes a buildfailure in glibc-initial, which doesn't build the linker.

+++ meta-linaro-toolchain/recipes-core/glibc/glibc_linaro-2.20.bb
@@ -196,7 +196,9 @@ do_compile () {
 # be fine.
 #
 do_install_append_aarch64-be() {
-    ln -sf ld-linux-aarch64_be.so.1 ${D}${base_libdir}/ld-linux-aarch64.so.1
+    if [ -e ${D}${base_libdir}/ld-linux-aarch64_be.so.1 ] ; then
+        ln -sf ld-linux-aarch64_be.so.1 ${D}${base_libdir}/ld-linux-aarch64.so.1
+    fi
 }
 
 require recipes-core/glibc/glibc-package.inc
